+++ TeamTalkClassic/gui/HostManagerDlg.cpp
@@ -142,7 +142,7 @@ void CHostManagerDlg::DisplayHosts()
     if(m_wndHosts.GetCount()==0)
         OnButtonNew();
 
-    for(int i=0;i<m_vecHosts.size();i++)
+    for(size_t i=0;i<m_vecHosts.size();i++)
         m_wndHostAddress.AddString( STR_UTF8(m_vecHosts[i].szAddress.c_str()));
 
     if(m_vecHosts.size()>0)
@@ -299,8 +299,8 @@ void CHostManagerDlg::OnSelchangeListHosts()
             int first = 0;
             while((m_wndHosts.GetItemData(first) & PUBSERVER_ITEMDATA) == 0)first++;
             index = index - first;
-            ASSERT(index < m_pubservers.size());
-            if(index < m_pubservers.size())
+            ASSERT(index < int(m_pubservers.size()));
+            if(index < int(m_pubservers.size()))
             {
                 entry = m_pubservers[index];
                 ok = TRUE;
@@ -444,7 +444,7 @@ void CHostManagerDlg::OnBnClickedButtonDelentry()
     int index = m_wndHostAddress.GetCurSel();
     if(index == CB_ERR)
         return;
-    if(index < m_vecHosts.size())
+    if(index < int(m_vecHosts.size()))
     {
         m_delHosts.push_back(m_vecHosts[index]);
         m_vecHosts.erase(m_vecHosts.begin()+index);
@@ -455,7 +455,7 @@ void CHostManagerDlg::OnBnClickedButtonDelentry()
 void CHostManagerDlg::OnCbnSelchangeComboHostaddress()
 {
     int index = m_wndHostAddress.GetCurSel();
-    if(index != CB_ERR && index<m_vecHosts.size())
+    if(index != CB_ERR && index < int(m_vecHosts.size()))
     {
         CString s;s.Format(_T("%d"),m_vecHosts[index].nTcpPort);
         m_wndHostPort.SetWindowText(s);
